:shower: cleaning

+++ src/flups.cpp
@@ -133,7 +133,7 @@ FLUPS_SIZE flups_get_allocSize(FLUPS_Solver* s){
     return(FLUPS_SIZE) s->get_allocSize();
 }
 
-void flups_get_spectralInfo(FLUPS_Solver* s, double kfact[3], double koffset[3], int symstart[3]){
+void flups_get_spectralInfo(FLUPS_Solver* s, double kfact[3], double koffset[3], double symstart[3]){
     s->get_spectralInfo(kfact,koffset,symstart);
 }
